requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
extra-streamlit-components==0.1.55
fastjsonschema==2.16.1
Flask==2.1.3
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
fonttools==4.34.4
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipwin==0.5.2
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
pyjsparser==2.7.1
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
PyPrind==2.11.3
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rich==12.5.1
scikit-learn==1.0.2
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
streamlit==1.11.0
streamlit-authenticator==0.2.1
streamlit-option-menu==0.3.2
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
zipp==3.8.1
